+++ packages/medusa/src/services/__tests__/order.js
@@ -25,6 +25,9 @@ describe("OrderService", () => {
     getSubtotal: o => {
       return o.subtotal || 0
     },
+    getPaidTotal: o => {
+      return o.paid_total || 0
+    },
   }
 
   const eventBusService = {
@@ -858,16 +861,20 @@ describe("OrderService", () => {
     })
   })
 
-  describe("receiveReturn", () => {
+  describe("registerReturnReceived", () => {
     const order = {
       items: [
         {
           id: "item_1",
           quantity: 10,
-          returned_quantity: 0,
+          returned_quantity: 10,
         },
       ],
-      payments: [{ id: "payment_test" }],
+      payments: [{ id: "payment_test", amount: 100 }],
+      refunded_total: 0,
+      paid_total: 100,
+      refundable_amount: 100,
+      total: 100,
     }
     const orderRepo = MockRepository({
       findOneWithRelations: (rel, q) => {
@@ -877,24 +884,6 @@ describe("OrderService", () => {
         }
       },
     })
-
-    const returnService = {
-      retrieve: () => {
-        return Promise.resolve({
-          order_id: IdMap.getId("order"),
-        })
-      },
-      receiveReturn: jest
-        .fn()
-        .mockImplementation((id, items, amount, mism) =>
-          id === IdMap.getId("good")
-            ? Promise.resolve({ items, status: "received", refund_amount: 100 })
-            : Promise.resolve({ status: "requires_action" })
-        ),
-      withTransaction: function() {
-        return this
-      },
-    }
 
     const paymentProviderService = {
       refundPayment: jest
@@ -907,20 +896,11 @@ describe("OrderService", () => {
       },
     }
 
-    const lineItemService = {
-      update: jest.fn(),
-      withTransaction: function() {
-        return this
-      },
-    }
-
     const orderService = new OrderService({
       manager: MockManager,
       orderRepository: orderRepo,
       paymentProviderService,
       totalsService,
-      returnService,
-      lineItemService,
       eventBusService,
     })
 
@@ -929,29 +909,11 @@ describe("OrderService", () => {
     })
 
     it("calls order model functions", async () => {
-      const items = [
-        {
-          item_id: "item_1",
-          quantity: 10,
-        },
-      ]
-      await orderService.receiveReturn(
-        IdMap.getId("order"),
-        IdMap.getId("good"),
-        items
-      )
-
-      expect(returnService.receiveReturn).toHaveBeenCalledTimes(1)
-      expect(returnService.receiveReturn).toHaveBeenCalledWith(
-        IdMap.getId("good"),
-        items,
-        undefined,
-        false
-      )
-
-      expect(lineItemService.update).toHaveBeenCalledTimes(1)
-      expect(lineItemService.update).toHaveBeenCalledWith("item_1", {
-        returned_quantity: 10,
+      await orderService.registerReturnReceived(IdMap.getId("order"), {
+        id: IdMap.getId("good"),
+        order_id: IdMap.getId("order"),
+        status: "received",
+        refund_amount: 100,
       })
 
       expect(orderRepo.save).toHaveBeenCalledTimes(1)
@@ -969,136 +931,23 @@ describe("OrderService", () => {
     })
 
     it("return with custom refund", async () => {
-      const items = [
-        {
-          item_id: IdMap.getId("existingLine"),
-          quantity: 10,
-        },
-      ]
-      await orderService.receiveReturn(
+      await orderService.registerReturnReceived(
         IdMap.getId("order"),
-        IdMap.getId("good"),
-        items,
-        102
-      )
-
-      expect(returnService.receiveReturn).toHaveBeenCalledTimes(1)
-      expect(returnService.receiveReturn).toHaveBeenCalledWith(
-        IdMap.getId("good"),
-        items,
-        102,
-        false
-      )
-    })
-
-    it("calls order model functions and sets partially_returned", async () => {
-      const items = [
         {
-          item_id: IdMap.getId("existingLine"),
-          quantity: 2,
+          id: IdMap.getId("good"),
+          order_id: IdMap.getId("order"),
+          status: "received",
+          refund_amount: 95,
         },
-      ]
-      await orderService.receiveReturn(
-        IdMap.getId("order"),
-        IdMap.getId("good"),
-        items
+        95
       )
 
-      expect(orderRepo.save).toHaveBeenCalledTimes(1)
-      expect(orderRepo.save).toHaveBeenCalledWith({
-        ...order,
-        fulfillment_status: "partially_returned",
-      })
-    })
-
-    it("sets requires_action on additional items", async () => {
-      await orderService.receiveReturn(
-        IdMap.getId("order"),
-        IdMap.getId("action"),
-        [
-          {
-            item_id: IdMap.getId("existingLine2"),
-            quantity: 2,
-          },
-        ]
+      expect(paymentProviderService.refundPayment).toHaveBeenCalledTimes(1)
+      expect(paymentProviderService.refundPayment).toHaveBeenCalledWith(
+        order.payments,
+        95,
+        "return"
       )
-
-      expect(orderRepo.save).toHaveBeenCalledTimes(1)
-      expect(orderRepo.save).toHaveBeenCalledWith({
-        ...order,
-        fulfillment_status: "requires_action",
-      })
-    })
-  })
-
-  describe("requestReturn", () => {
-    const order = {
-      items: [
-        {
-          id: "item_1",
-          quantity: 10,
-          returned_quantity: 0,
-        },
-      ],
-      payments: [{ id: "payment_test" }],
-    }
-    const orderRepo = MockRepository({
-      findOneWithRelations: (rel, q) => {
-        switch (q.where.id) {
-          default:
-            return Promise.resolve(order)
-        }
-      },
-    })
-
-    const returnService = {
-      create: jest.fn(() => Promise.resolve({ id: "ret" })),
-      fulfill: jest.fn(() => Promise.resolve({ id: "ret" })),
-      withTransaction: function() {
-        return this
-      },
-    }
-
-    const orderService = new OrderService({
-      manager: MockManager,
-      orderRepository: orderRepo,
-      totalsService,
-      returnService,
-      eventBusService,
-    })
-
-    beforeEach(async () => {
-      jest.clearAllMocks()
-    })
-
-    it("successfully creates return request", async () => {
-      const items = [
-        {
-          item_id: IdMap.getId("existingLine"),
-          quantity: 10,
-        },
-      ]
-      const shipping_method = {
-        id: IdMap.getId("return-shipping"),
-        price: 2,
-      }
-      await orderService.requestReturn(
-        "processed-order",
-        items,
-        shipping_method
-      )
-
-      expect(returnService.create).toHaveBeenCalledTimes(1)
-      expect(returnService.create).toHaveBeenCalledWith(
-        {
-          items,
-          shipping_method,
-          refund_amount: undefined,
-          order_id: "processed-order",
-        },
-        order
-      )
-      expect(returnService.fulfill).toHaveBeenCalledWith("ret")
     })
   })
 
@@ -1201,125 +1050,6 @@ describe("OrderService", () => {
     })
   })
 
-  describe("registerSwapReceived", () => {
-    beforeEach(async () => {
-      jest.clearAllMocks()
-    })
-    const orderRepo = MockRepository({
-      findOneWithRelations: () => Promise.resolve({ id: IdMap.getId("order") }),
-    })
-
-    it("fails if order/swap relationship not satisfied", async () => {
-      const swapService = {
-        retrieve: jest
-          .fn()
-          .mockReturnValue(
-            Promise.resolve({ id: "1235", order_id: IdMap.getId("order_1") })
-          ),
-        withTransaction: function() {
-          return this
-        },
-      }
-
-      const orderService = new OrderService({
-        manager: MockManager,
-        orderRepository: orderRepo,
-        totalsService,
-        swapService,
-        eventBusService,
-      })
-
-      const res = orderService.registerSwapReceived(
-        IdMap.getId("order"),
-        "1235"
-      )
-      await expect(res).rejects.toThrow("Swap must belong to the given order")
-    })
-
-    it("fails if swap doesn't have status received", async () => {
-      const swapService = {
-        retrieve: jest.fn().mockReturnValue(
-          Promise.resolve({
-            id: "1235",
-            order_id: IdMap.getId("order"),
-            return_order: { status: "requested" },
-          })
-        ),
-        withTransaction: function() {
-          return this
-        },
-      }
-      const orderService = new OrderService({
-        manager: MockManager,
-        orderRepository: orderRepo,
-        swapService,
-        totalsService,
-        eventBusService: { emit: jest.fn().mockReturnValue(Promise.resolve()) },
-      })
-
-      const res = orderService.registerSwapReceived(
-        IdMap.getId("order"),
-        "1235"
-      )
-      await expect(res).rejects.toThrow("Swap is not received")
-    })
-
-    it("registers a swap as received", async () => {
-      const orderRepo = MockRepository({
-        findOneWithRelations: () =>
-          Promise.resolve({
-            id: IdMap.getId("order_123"),
-            items: [
-              {
-                id: IdMap.getId("1234"),
-                returned_quantity: 0,
-                quantity: 1,
-              },
-            ],
-          }),
-      })
-
-      const swapService = {
-        retrieve: jest.fn().mockReturnValue(
-          Promise.resolve({
-            id: "1235",
-            order_id: IdMap.getId("order_123"),
-            return_order: {
-              status: "received",
-              items: [{ item_id: IdMap.getId("1234"), quantity: 1 }],
-            },
-          })
-        ),
-        withTransaction: function() {
-          return this
-        },
-      }
-
-      const lineItemService = {
-        update: jest.fn(),
-        withTransaction: function() {
-          return this
-        },
-      }
-
-      const orderService = new OrderService({
-        manager: MockManager,
-        orderRepository: orderRepo,
-        totalsService,
-        swapService,
-        lineItemService,
-        eventBusService,
-      })
-
-      await orderService.registerSwapReceived(IdMap.getId("order_123"), "1235")
-
-      expect(lineItemService.update).toHaveBeenCalledTimes(1)
-      expect(lineItemService.update).toHaveBeenCalledWith(IdMap.getId("1234"), {
-        returned_quantity: 1,
-      })
-    })
-  })
-
   describe("createRefund", () => {
     beforeEach(async () => {
       jest.clearAllMocks()
@@ -1341,13 +1071,15 @@ describe("OrderService", () => {
         }
 
         return Promise.resolve({
-          id: IdMap.getId("order"),
+          id: IdMap.getId("order_123"),
           payments: [
             {
               id: "payment",
             },
           ],
           total: 100,
+          paid_total: 100,
+          refundable_amount: 100,
           refunded_total: 0,
         })
       },