+++ packages/medusa-plugin-discount-generator/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.2](https://github.com/medusajs/medusa/compare/medusa-plugin-discount-generator@1.1.1...medusa-plugin-discount-generator@1.1.2) (2021-04-20)
+
+**Note:** Version bump only for package medusa-plugin-discount-generator
+
+
+
+
+
+## [1.1.1](https://github.com/medusajs/medusa/compare/medusa-plugin-discount-generator@1.1.0...medusa-plugin-discount-generator@1.1.1) (2021-04-20)
+
+**Note:** Version bump only for package medusa-plugin-discount-generator
+
+
+
+
+
 # 1.1.0 (2021-01-26)